refactor(format): split analysis html renderer
