+++ crates/tokmd-format/src/analysis/html/format.rs
@@ -1,0 +1,24 @@
+//! Shared HTML formatting helpers.
+
+pub(super) fn format_number(n: usize) -> String {
+    if n >= 1_000_000 {
+        format!("{:.1}M", n as f64 / 1_000_000.0)
+    } else if n >= 1_000 {
+        format!("{:.1}K", n as f64 / 1_000.0)
+    } else {
+        n.to_string()
+    }
+}
+
+pub(super) fn format_pct(ratio: f64) -> String {
+    format!("{:.1}%", ratio * 100.0)
+}
+
+pub(super) fn escape_html(value: &str) -> String {
+    value
+        .replace('&', "&amp;")
+        .replace('<', "&lt;")
+        .replace('>', "&gt;")
+        .replace('"', "&quot;")
+        .replace('\'', "&#x27;")
+}
